fix(discord): 401 bug
